+++ bin/sidekiqload
@@ -8,6 +8,10 @@ require 'net/http'
 require "sidekiq"
 Sidekiq.redis { |c| c.flushall }
 
+CONCURRENCY = 10
+ITERATIONS = 1
+BATCH = 10_000
+
 class LoadJob
   include Sidekiq::Job
   sidekiq_options retry: 1
@@ -26,18 +30,22 @@ class LoadJob
     puts "Request Error!!!" unless res.is_a?(Net::HTTPSuccess)
   end
   
+  # 1000
+  # 10 conc 6.60s 82636 RSS 96.33 CPU
+  # 10_000
+  # 10 conc 41.49 91408 RSS 102.72 CPU
   def perform(idx)
     start = Process.clock_gettime(Process::CLOCK_MONOTONIC)
-    sample_get_request # 0.91
-    fib(33) # 0.36 71% IO, 29% CPU
-    puts "#{idx} ended #{Process.clock_gettime(Process::CLOCK_MONOTONIC) - start}\n"
+    fib(18) # 0.000046
+    File.open("tmp/log.txt", "a") { |f| f.write "#{idx} ended #{Process.clock_gettime(Process::CLOCK_MONOTONIC) - start}\n" }
   end
 end
 
 def get_rss_cpu(pid, cpus = nil)
   data = `ps -o rss,%cpu= -p #{pid}`.split("\n").last.strip.split(" ")
-  cpus << data[1].to_f if cpus
-  "RSS: #{data[0]} %CPU: #{data[1]}"
+  cpu = data[1].to_f.round(2)
+  cpus << cpu if !cpus.nil? && cpu > 0
+  "RSS: #{data[0]} %CPU: #{cpu}"
 end
 
 def jobs_count
@@ -48,13 +56,11 @@ end
 
 # when trying to upload 1_000_000 at once
 # ruby_connection/buffered_io.rb:140:in `block in fill_buffer': RedisClient::ReadTimeoutError (RedisClient::ReadTimeoutError)
-iter = 1
-pre_count = 200
-count = iter * pre_count
+count = ITERATIONS * BATCH
 
 Sidekiq.logger.error "Uploading #{count} jobs..."
-iter.times do |i|
-  arr = Array.new(pre_count) { |idx| [idx] }
+ITERATIONS.times do |i|
+  arr = Array.new(BATCH) { |idx| [idx] }
   Sidekiq::Client.push_bulk("class" => LoadJob, "args" => arr)
 end
 Sidekiq.logger.error "Created #{count} jobs"
@@ -73,7 +79,7 @@ p2 = fork do
   
   begin
     cli = Sidekiq::CLI.instance
-    cli.parse ["-r", "tmp/temp.rb", "-c", "10"] # require an empty file to pass validation
+    cli.parse ["-r", "tmp/temp.rb", "-t", "100", "-c", "#{CONCURRENCY}"] # require an empty file to pass validation
     Sidekiq.logger.error "Concurrency: #{Sidekiq.default_configuration.concurrency}"
   
     cli.run(boot_app: false)
@@ -84,25 +90,22 @@ p2 = fork do
 end
 
 start = Time.now
-
-total = jobs_count
 cpus = []
-Sidekiq.logger.error("#{get_rss_cpu(p2, cpus)} Pending: #{total}")
 loop do
-  sleep 2.0
   total = jobs_count
-  Sidekiq.logger.error("#{get_rss_cpu(p2, cpus)} Pending: #{total}")
+  Sidekiq.logger.error("#{get_rss_cpu(p2, cpus)} Pending: #{total} Time: #{(Time.now - start).round(2)} sec")
   if total == 0
-    Sidekiq.logger.error("#{get_rss_cpu(p2, cpus)} Pending: #{total}")
     Sidekiq.logger.error("Done, #{count} jobs in #{Time.now - start} sec")
     break
   end
+  sleep 2
 end
+Sidekiq.logger.error("Sending INT...")
 Process.kill("INT", p2)
-3.times do
-  sleep(2)
-  Sidekiq.logger.error("Finalizing... #{get_rss_cpu(p2)}")
-end
+# 5.times do
+#   sleep(2)
+#   Sidekiq.logger.error("Finalizing... #{get_rss_cpu(p2, cpus)}")
+# end
 Process.wait(p2)
 Sidekiq.logger.error("Real time is: #{Process.clock_gettime(Process::CLOCK_MONOTONIC) - real_start}")
 Sidekiq.logger.error("Average CPU usage: #{cpus.sum / cpus.size}")
