+++ lib/jiggler/manager.rb
@@ -38,10 +38,10 @@ module Jiggler
     private
 
     def wait_for_workers
-      logger.info('Waiting for workers to finish...')
+      logger.warn('Waiting for workers to finish...')
       @workers.each(&:wait)
       @shutdown_task.stop
-      logger.info('All workers finished')
+      logger.warn('All workers finished')
     end
 
     def schedule_shutdown
