+++ lib/jiggler/worker.rb
@@ -22,18 +22,23 @@ module Jiggler
       @runner = safe_async('Worker') do
         @tid = tid
         loop do
-          break @callback.call(self) if @done
+          if @done
+            @runner = nil
+            break @callback.call(self)
+          end
+
           process_job
 
           # pass control to other fibers
           sleep(0)
-          # sleep(0) appears to work slower than Async::Task.current.yield
-          # but it seems to be more reliable - at least looks like
-          # Monitor & Poller are getting more runs with it in CPU-heavy load
+          # sleep(0) seems to be more reliable than Async::Task.current.yield
+          # at least looks like Monitor & Poller are getting more runs with it in CPU-heavy load
         rescue Async::Stop
+          @runner = nil
           break @callback.call(self)
         rescue => err
           collection.incr_failures
+          @runner = nil
           break @callback.call(self, err)     
         end
       end
@@ -49,7 +54,9 @@ module Jiggler
     end
 
     def wait
+      # logger.warn("worker #{@tid} is waiting...")
       @runner&.wait
+      # logger.warn("worker #{@tid} done")
     end
 
     private
@@ -63,13 +70,13 @@ module Jiggler
 
     def fetch_one
       queue, args = config.with_sync_redis { |conn| conn.blocking_call(false, 'BRPOP', *queues, TIMEOUT) }
-      if queue
-        if @done
-          requeue(queue, args)
-          nil
-        else
-          CurrentJob.new(queue: queue, args: args)
-        end
+      return nil unless queue
+
+      if @done
+        requeue(queue, args)
+        nil
+      else
+        CurrentJob.new(queue: queue, args: args)
       end
     rescue Async::Stop => err
       raise err
@@ -77,49 +84,50 @@ module Jiggler
       # error_message='undefined method `zero?' for nil:NilClass' context='Fetch error' tid=1tpj
       # sometimes happens in async-pool-0.3.12/lib/async/pool/controller.rb:213:in `reuse'
       handle_fetch_error(err)
+      nil
     end
     
     def execute_job
-      parsed_args = JSON.parse(current_job.args)
-      begin
-        execute(parsed_args, current_job.queue)
-        collection.incr_processed
-      rescue Async::Stop => err
-        raise err
-      rescue UnknownJobError => err
-        handle_exception(
-          err,
-          {
-            error_class: err.class.name,
-            job: parsed_args,
-            tid: @tid
-          },
-          raise_ex: true
-        )
-      rescue Exception => ex
-        handle_exception(
-          ex,
-          {
-            context: '\'Internal exception\'',
-            tid: @tid,
-            jid: parsed_args['jid']
-          },
-          raise_ex: true
-        )
-      end
+      parsed_args = Oj.load(current_job.args, mode: :compat)
+      execute(parsed_args, current_job.queue)
+    rescue Async::Stop => err
+      raise err
+    rescue UnknownJobError => err
+      collection.incr_failures
+      handle_exception(
+        err,
+        {
+          error_class: err.class.name,
+          job: parsed_args,
+          tid: @tid
+        },
+        raise_ex: false
+      )
     rescue JSON::ParserError => err
       collection.incr_failures
       logger.error('Worker') { "Failed to parse job: #{current_job.args}" }
+    rescue Exception => ex
+      handle_exception(
+        ex,
+        {
+          context: '\'Internal exception\'',
+          tid: @tid,
+          jid: parsed_args['jid']
+        },
+        raise_ex: true
+      )
     end
 
     def execute(parsed_job, queue)
-      klass = collection.fetch_job_class(parsed_job['name'])
-      instance = klass.new
+      instance = constantize(parsed_job['name']).new
+      add_current_job_to_collection(parsed_job, parsed_job['queue'])
 
-      add_current_job_to_collection(parsed_job, klass.queue)
       retrier.wrapped(instance, parsed_job, queue) do
         instance.perform(*parsed_job['args'])
       end
+
+      collection.incr_processed
+      # logger.warn("worker #{@tid} has done job")
     ensure
       remove_current_job_from_collection
     end
@@ -159,6 +167,19 @@ module Jiggler
 
     def queues
       @queues ||= config.prefixed_queues
+    end
+
+    def constantize(str)
+      return Object.const_get(str) unless str.include?('::')
+
+      names = str.split('::')
+      names.shift if names.empty? || names.first.empty?
+
+      names.inject(Object) do |constant, name|
+        constant.const_get(name, false)
+      end
+    rescue => err
+      raise UnknownJobError, 'Cannot initialize job'
     end
   end
 end
